ML/Feature8/requirements.txt
torch
transformers
sentence-transformers
datasets
scikit-learn
numpy
pandas
faiss-cpu
spacy
pyyaml
